Adding sparse matrix

+++ cpp/lib/Mrpt_Sparse.h
@@ -1907,7 +1907,7 @@ private:
       if (sparse_input){
         Q.setConstant(0.0);
         for (int k = 0; k < X_Sparse.outerSize(); ++k) {
-          for (Eigen::SparseMatrix<double>::InnerIterator it(X_Sparse, k); it; ++it) {
+          for (Eigen::SparseMatrix<float>::InnerIterator it(X_Sparse, k); it; ++it) {
             if (it.col() == indices[i]) {
               Q(it.row(), i) = it.value();
             }
